+++ source/renderer/modules/markdown-editor/modes/multiplex.js
@@ -203,6 +203,158 @@ const highlightingModes = {
 })(function (CodeMirror) {
   'use strict'
 
+  /*  This function is a copy of CodeMirror.multiplexingMode with a small modification made to
+      the token function.  CodeMirror's multiplexing mode addon involves a brute force check of
+      every internal mode object, and when these checks are regular expressions on long lines
+      they can take a very long amount of time to complete.
+
+      However, Zettlr's use case is unique, because all but a few of the modes have an opening
+      regular expression that contains triple backticks (```).  This modified multiplexer takes
+      advantage of that commonality to eliminate the bulk of the checks.
+  */
+  let zettlrMultiplexer = function(outer /*, others */) {
+    // Others should be {open, close, mode [, delimStyle] [, innerStyle]} objects
+    var others = Array.prototype.slice.call(arguments, 1);
+  
+    function indexOf(string, pattern, from, returnEnd) {
+      if (typeof pattern == "string") {
+        var found = string.indexOf(pattern, from);
+        return returnEnd && found > -1 ? found + pattern.length : found;
+      }
+      var m = pattern.exec(from ? string.slice(from) : string);
+      return m ? m.index + from + (returnEnd ? m[0].length : 0) : -1;
+    }
+  
+    return {
+      startState: function() {
+        return {
+          outer: CodeMirror.startState(outer),
+          innerActive: null,
+          inner: null
+        };
+      },
+  
+      copyState: function(state) {
+        return {
+          outer: CodeMirror.copyState(outer, state.outer),
+          innerActive: state.innerActive,
+          inner: state.innerActive && CodeMirror.copyState(state.innerActive.mode, state.inner)
+        };
+      },
+  
+      /*  This token function has been modified from the version that can be found at
+          https://github.com/codemirror/CodeMirror/blob/master/addon/mode/multiplex.js
+          by performing a check to see if the stream content (the line being tokenized) 
+          contains three backticks in a row.  If the line does not contain these backticks
+          we can guarantee that all of the syntax highlighting regular expressions which
+          require `{3} will fail, and so there is no need to perform the expensive check
+          on them.  We know that indexOf will return -1.
+
+          The only other component is to know which modes can take advantage of the 
+          backtick knowledge.  We mark these modes with a .backTicks property.
+      */
+      token: function(stream, state) {
+        if (!state.innerActive) {
+          var cutOff = Infinity, oldContent = stream.string;
+
+          // Check if the stream contains triple backticks ```
+          let hasBackticks = oldContent.includes("```")
+
+          for (var i = 0; i < others.length; ++i) {
+            var other = others[i];
+
+            // If we do not have the triple backticks in this line AND the mode requires the
+            // triple backticks in its open pattern, we know that indexOf will return -1. 
+            // Because stream.pos does not appear to ever naturally take the value of -1, none 
+            // of the case statements inside the rest of the for loop will be triggered. Thus
+            // we can safely abort this cycle of the loop.
+            if (!hasBackticks && other.backTicks) {
+              continue
+            }
+
+            var found = indexOf(oldContent, other.open, stream.pos);
+            if (found == stream.pos) {
+              if (!other.parseDelimiters) stream.match(other.open);
+              state.innerActive = other;
+  
+              // Get the outer indent, making sure to handle CodeMirror.Pass
+              var outerIndent = 0;
+              if (outer.indent) {
+                var possibleOuterIndent = outer.indent(state.outer, "", "");
+                if (possibleOuterIndent !== CodeMirror.Pass) outerIndent = possibleOuterIndent;
+              }
+  
+              state.inner = CodeMirror.startState(other.mode, outerIndent);
+              return other.delimStyle && (other.delimStyle + " " + other.delimStyle + "-open");
+            } else if (found != -1 && found < cutOff) {
+              cutOff = found;
+            }
+          }
+
+          if (cutOff != Infinity) stream.string = oldContent.slice(0, cutOff);
+          var outerToken = outer.token(stream, state.outer);
+          if (cutOff != Infinity) stream.string = oldContent;
+          return outerToken;
+        } else {
+          var curInner = state.innerActive, oldContent = stream.string;
+          if (!curInner.close && stream.sol()) {
+            state.innerActive = state.inner = null;
+            return this.token(stream, state);
+          }
+          var found = curInner.close ? indexOf(oldContent, curInner.close, stream.pos, curInner.parseDelimiters) : -1;
+          if (found == stream.pos && !curInner.parseDelimiters) {
+            stream.match(curInner.close);
+            state.innerActive = state.inner = null;
+            return curInner.delimStyle && (curInner.delimStyle + " " + curInner.delimStyle + "-close");
+          }
+          if (found > -1) stream.string = oldContent.slice(0, found);
+          var innerToken = curInner.mode.token(stream, state.inner);
+          if (found > -1) stream.string = oldContent;
+  
+          if (found == stream.pos && curInner.parseDelimiters)
+            state.innerActive = state.inner = null;
+  
+          if (curInner.innerStyle) {
+            if (innerToken) innerToken = innerToken + " " + curInner.innerStyle;
+            else innerToken = curInner.innerStyle;
+          }
+  
+          return innerToken;
+        }
+      },
+  
+      indent: function(state, textAfter, line) {
+        var mode = state.innerActive ? state.innerActive.mode : outer;
+        if (!mode.indent) return CodeMirror.Pass;
+        return mode.indent(state.innerActive ? state.inner : state.outer, textAfter, line);
+      },
+  
+      blankLine: function(state) {
+        var mode = state.innerActive ? state.innerActive.mode : outer;
+        if (mode.blankLine) {
+          mode.blankLine(state.innerActive ? state.inner : state.outer);
+        }
+        if (!state.innerActive) {
+          for (var i = 0; i < others.length; ++i) {
+            var other = others[i];
+            if (other.open === "\n") {
+              state.innerActive = other;
+              state.inner = CodeMirror.startState(other.mode, mode.indent ? mode.indent(state.outer, "", "") : 0);
+            }
+          }
+        } else if (state.innerActive.close === "\n") {
+          state.innerActive = state.inner = null;
+        }
+      },
+  
+      electricChars: outer.electricChars,
+  
+      innerMode: function(state) {
+        return state.inner ? {state: state.inner, mode: state.innerActive.mode} : {state: state.outer, mode: outer};
+      }
+    };
+  };
+
   /**
   * MULTIPLEX MODE: This will by default load our internal mode cascade
   * (consisting of the zkn-mode, the spellchecker and finally the gfm
@@ -221,11 +373,15 @@ const highlightingModes = {
         close: /`{3}|~{3}/,
         mode: CodeMirror.getMode(config, mimeType),
         delimStyle: 'formatting-code-block',
-        innerStyle: 'fenced-code'
+        innerStyle: 'fenced-code',
+
+        // To take advantage of the modified multiplexer, we add a property here to mark that this 
+        // opening regex will not be necessary if there are no triple backticks in the line being checked
+        backTicks: true   
       })
     }
 
-    return CodeMirror.multiplexingMode(
+    return zettlrMultiplexer(
       CodeMirror.getMode(config, 'spellchecker'), // Default mode
       ...codeModes,
       {
@@ -233,7 +389,11 @@ const highlightingModes = {
         close: /`{3}|~{3}/,
         mode: CodeMirror.getMode(config, 'text/plain'),
         delimStyle: 'formatting-code-block',
-        innerStyle: 'fenced-code'
+        innerStyle: 'fenced-code',
+
+        // To take advantage of the modified multiplexer, we add a property here to mark that this 
+        // opening regex will not be necessary if there are no triple backticks in the line being checked
+        backTicks: true
       },
       {
         open: /\$\$\s*$/,